requirements.txt
psycopg2-binary
requests
schedule
matplotlib
pandas
scikit-image
geopandas
contextily